+++ src/elements/Fragment.php
@@ -7,6 +7,9 @@ use craft\base\Element;
 use craft\elements\actions\Delete;
 use craft\elements\actions\Restore;
 use craft\elements\actions\SetStatus;
+use craft\elements\conditions\ElementConditionInterface;
+use craft\elements\conditions\entries\EntryCondition;
+use craft\elements\conditions\users\UserCondition;
 use craft\elements\db\ElementQueryInterface;
 use craft\elements\User;
 use craft\events\DefineHtmlEvent;
@@ -38,24 +41,11 @@ class Fragment extends Element
     public ?int $fragmentTypeId;
     public ?int $zoneId;
     public ?array $settings = [];
+    public ElementConditionInterface|null $_entryCondition = null;
+    public ElementConditionInterface|null $_userCondition = null;
 
     public function __construct($config = [])
     {
-        if (isset($config['settings'])) {
-            $config['settings'] = Json::decode($config['settings']);
-        } else {
-            $config['settings'] = [
-                'visibility' => [
-                    'ruletype' => '',
-                    'rules' => [],
-                ],
-            ];
-        }
-
-        if (!is_array($config['settings']['visibility']['rules'])) {
-            $config['settings']['visibility']['rules'] = [];
-        }
-
         parent::__construct($config);
     }
 
@@ -350,7 +340,8 @@ class Fragment extends Element
 
             $record->zoneId = (int)$this->zoneId;
             $record->fragmentTypeId = (int)$this->fragmentTypeId;
-            $record->settings = Json::encode($this->settings);
+            $record->entryCondition = $this->getEntryCondition()->getConfig();
+            $record->userCondition = $this->getUserCondition()->getConfig();
             $record->save(false);
 
             if (!$this->duplicateOf && $isNew) {
@@ -508,6 +499,60 @@ class Fragment extends Element
         }
 
         return parent::tableAttributeHtml($attribute);
+    }
+
+    public function getEntryCondition(): ElementConditionInterface
+    {
+        $condition = $this->_entryCondition ?? new EntryCondition();
+        $condition->mainTag = 'div';
+        $condition->name = 'entryCondition';
+
+        return $condition;
+    }
+
+    /**
+     * @throws InvalidConfigException
+     */
+    public function setEntryCondition(ElementConditionInterface|string|array|null $condition): void
+    {
+        if (is_string($condition)) {
+            $condition = Json::decodeIfJson($condition);
+        }
+
+        if (!$condition instanceof ElementConditionInterface) {
+            $condition['class'] = EntryCondition::class;
+            $condition = Craft::$app->getConditions()->createCondition($condition);
+        }
+        $condition->forProjectConfig = false;
+
+        $this->_entryCondition = $condition;
+    }
+
+    public function getUserCondition(): ElementConditionInterface
+    {
+        $condition = $this->_userCondition ?? new UserCondition();
+        $condition->mainTag = 'div';
+        $condition->name = 'userCondition';
+
+        return $condition;
+    }
+
+    /**
+     * @throws InvalidConfigException
+     */
+    public function setUserCondition(ElementConditionInterface|string|array|null $condition): void
+    {
+        if (is_string($condition)) {
+            $condition = Json::decodeIfJson($condition);
+        }
+
+        if (!$condition instanceof ElementConditionInterface) {
+            $condition['class'] = UserCondition::class;
+            $condition = Craft::$app->getConditions()->createCondition($condition);
+        }
+        $condition->forProjectConfig = false;
+
+        $this->_userCondition = $condition;
     }
 
     public static function gqlTypeNameByContext(mixed $context): string